requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
langdetect>=1.0.9
deep-translator>=1.11.0
gTTS>=2.4.0
python-docx>=1.1.0
pdfminer.six>=20221105
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-pptx>=1.0.0
Werkzeug>=2.3.0